[Scramble Input] This script scrambles the lines in input file and puts it in an output file
This was written for creating a certain level during Hack The Box. The script,
essentially reads the contents of $INFILE, and places each line at a random
location in $TEXTFILE. $TEXTFILE should contain a large amount of text for this
to work efficiently.

The long statement on Line 12, first shuffles the lines of text in $INFILE. Then
for each line, the awk statement computes a random number m and places that line
m lines below the current working position in $TEXTFILE. Since the initial lines were
shuffled, the output has the lines in a completely random order.

Lines 13-18 are a validity check. They ensure that the random number generated
did not overflow the total number of lines in $TEXTFILE, hence ensuring that all
lines from $INFILE are now placed in $TEXTFILE.

This script isn't perfect and all sorts of error checking could be implemented.
The output is horrible as well. But it did it's work and that's all I care