[cyberoam.sh] Support command line config option
This introduces a very important change to the script. Now one can
use the -c CONFPATH switch to specify the path of the configuration
file. With this change, it is now very easy to set cyberoam.sh as
a startup script for all users of a system with multiple users
sharing the same configuration file. A systemd service file will be
commited soon.