[gen-sigs.sh] This scripts generates a bunch of random 16-bit signatures
This script was written to automate the generation of a character array in C for
an assignment that required the simulation of a door locking system. The script
generates random characters for signatures and attaches them to a random 3-digit
number that represents a person. A C-style character array is also printed that
can simply be copied into the code without modifications.