New script, cont... algs4_dev.sh
This is the more important script of the two. It sets up the complete
development environment for the course.

The script optionally accepts a parameter for the project root dir. If the
parameter is gven the system will be setup to use the provided root directory,
else the default dir of ~/algs4/ will be used.